+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   of when the service runs #96
 - refactor(vscode): add name of the vscode instance to the sub heading of the vscode
   menu when it is running
+- fix(vscode): set a timeout for vscode commands - closes #101
 
 ## Version 0.14.1